+++ src/__/arrays/dropRightWhile.php
@@ -87,6 +87,8 @@ function dropArrayRightWhile($input, $comparison)
  * @see dropRight
  * @see dropRightWhile
  *
+ * @since 0.2.3 added to Bottomline
+ *
  * @return array|\Generator An array containing a subset of the input array with front items matching the condition
  * removed. If the provided iterable is not an array, then a Generator will be returned.
  */